Add app.rb and io_handler.rb
