+++ books.json
@@ -1,0 +1,3 @@
+[
+  "#<Book:0x00000178c0b36480>"
+]
