+++ people.json
@@ -1,0 +1,3 @@
+[
+  "#<Student:0x000002b75f332340>"
+]
